requirements.txt
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
timm>=0.9.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
PyYAML>=6.0
tensorboard>=2.13.0
einops>=0.7.0
geoopt>=0.5.0
Pillow>=9.0.0
opencv-python>=4.7.0